requirements.txt
python-dotenv==0.21.0
psycopg2-binary==2.9.6
numpy==1.23.5
pandas==1.5.3
flask==3.1.0
psycopg==3.2.6
psycopg2==2.9.10
openai==0.28
